requirements.txt
autopep8==1.6.0
flake8==4.0.1
Flask==2.0.3
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
SQLAlchemy==1.4.31
